Introducing Atomic property
